+++ chaoslib/run.py
@@ -16,7 +16,7 @@ from logzero import logger
 
 from chaoslib import __version__, substitute
 from chaoslib.activity import run_activities
-from chaoslib.configuration import load_configuration
+from chaoslib.configuration import load_configuration, load_dynamic_configuration
 from chaoslib.control import (
     Control,
     cleanup_controls,
@@ -278,6 +278,7 @@ class Runner:
         self.secrets = load_secrets(
             experiment.get("secrets", {}), self.config, secret_vars
         )
+        self.config = load_dynamic_configuration(self.config, self.secrets)
 
     def cleanup(self):
         pass