+++ CHANGELOG.md
@@ -10,6 +10,13 @@
 
 ### Added
 
+* `configuration.py` now have the ability to have load a dynamic configurations.
+  The dynamic config is a regular activity (probe) in the configuration section.
+  If there's a use-case for setting a configuration dynamically right before the experiment is starting.
+  It executes the probe,
+  and then the return value of this probe will be the config you wish to set.
+### Added
+
 * Configuration now supports providing a type to convert Environment Variables
   to a provided type. Addressing the issue [here](https://github.com/chaostoolkit/chaostoolkit/issues/254#issue-1062797767)
   To specifically convert an Environment Variable loaded into configuration to